Share one PerpetualPositionData aggregate implementation
The mapping from PerpetualPositionData to PerpetualPositionDataAggregate
was written out three times: in the positions coordinator, in the position
details coordinator, and in AutocloseUIModelFactory.

Move it to a single PerpetualPositionDataAggregateImpl in :gemcore, next to
the interface it satisfies. The details aggregate now delegates the base
members to it and keeps only the details-specific ones.
