+++ android/data/coordinators/src/main/kotlin/com/gemwallet/android/data/coordinators/perpetuals/GetPerpetualPositionImpl.kt
@@ -2,15 +2,14 @@ package com.gemwallet.android.data.coordinators.perpetuals
 
 import com.gemwallet.android.application.perpetual.coordinators.GetPerpetualPosition
 import com.gemwallet.android.data.repositories.perpetual.PerpetualRepository
+import com.gemwallet.android.domains.perpetual.aggregates.PerpetualPositionDataAggregate
+import com.gemwallet.android.domains.perpetual.aggregates.PerpetualPositionDataAggregateImpl
 import com.gemwallet.android.domains.perpetual.aggregates.PerpetualPositionDetailsDataAggregate
-import com.gemwallet.android.domains.perpetual.formatPnlWithPercentage
 import com.gemwallet.android.domains.price.ValueDirection
 import com.gemwallet.android.domains.price.toValueDirection
 import com.gemwallet.android.model.CurrencyFormatter
 import com.gemwallet.android.model.PriceChangeFormatter
-import com.wallet.core.primitives.Asset
 import com.wallet.core.primitives.Currency
-import com.wallet.core.primitives.PerpetualDirection
 import com.wallet.core.primitives.PerpetualId
 import com.wallet.core.primitives.PerpetualMarginType
 import com.wallet.core.primitives.PerpetualPositionData
@@ -28,7 +27,9 @@ class GetPerpetualPositionImpl @Inject constructor(
 
 class PerpetualPositionDetailsDataAggregateImpl(
     private val data: PerpetualPositionData,
-) : PerpetualPositionDetailsDataAggregate {
+) : PerpetualPositionDetailsDataAggregate,
+    PerpetualPositionDataAggregate by PerpetualPositionDataAggregateImpl(data) {
+
     private val amountFormatter = CurrencyFormatter(type = CurrencyFormatter.Type.Fiat, currency = Currency.USD)
     private val priceFormatter = CurrencyFormatter(currency = Currency.USD)
 
@@ -51,19 +52,7 @@ class PerpetualPositionDetailsDataAggregateImpl(
 
     override val fundingPaymentsDirection: ValueDirection = fundingPaymentsValue.toValueDirection()
 
-    override val positionId: String = data.position.id
-
     override val perpetualId: PerpetualId = data.position.perpetualId
-
-    override val asset: Asset = data.asset
-
-    override val name: String = data.perpetual.name
-
-    override val direction: PerpetualDirection = data.position.direction
-
-    override val leverage: Int = data.position.leverage.toInt()
-
-    override val marginAmount: String = amountFormatter.string(data.position.marginAmount)
 
     override val entryValue: Double? = data.position.entryPrice
 
@@ -72,10 +61,4 @@ class PerpetualPositionDetailsDataAggregateImpl(
     override val stopLoss: Double? = data.position.stopLoss?.price
 
     override val takeProfit: Double? = data.position.takeProfit?.price
-
-    override val pnlWithPercentage: String
-        get() = formatPnlWithPercentage(data.position.pnl, data.position.marginAmount)
-
-    override val pnlState: ValueDirection
-        get() = data.position.pnl.toValueDirection()
 }
